requirements.txt
pytest
coverage
pydoc-markdown
licenseheaders

future
msgpack
braceexpand
simplejson
pyzmq
